(ref) move to_0x_hex to utils. add tests
