+++ hummingbot/connector/exchange/derive/derive_auth.py
@@ -8,7 +8,7 @@ from lyra_v2_action_signing import SignedAction, TradeModuleData, utils
 from web3 import Web3
 
 from hummingbot.connector.exchange.derive import derive_constants as CONSTANTS, derive_web_utils as web_utils
-from hummingbot.connector.other.to_0x_hex import to_0x_hex
+from hummingbot.connector.utils import to_0x_hex
 from hummingbot.core.web_assistant.auth import AuthBase
 from hummingbot.core.web_assistant.connections.data_types import RESTMethod, RESTRequest, WSRequest
